feat(#636): adapt dialog wrapper
