+++ app/lib/common/widgets/dialog_action.dart
@@ -2,35 +2,6 @@ import 'package:flutter/cupertino.dart';
 
 import '../module.dart';
 
-class AdaptiveDialogWrapper extends StatelessWidget {
-  const AdaptiveDialogWrapper({
-    super.key,
-    required this.title,
-    required this.content,
-    required this.actions,
-  });
-
-  final String title;
-  final Widget? content;
-  final List<DialogAction> actions;
-
-  @override
-  Widget build(BuildContext context) {
-    final materialContent = getPlatform() == SupportedPlatform.ios
-      ? Card(
-          color: Colors.transparent,
-          elevation: 0,
-          child: content,
-        )
-      : content;
-    return AlertDialog.adaptive(
-      title: Text(title),
-      content: materialContent,
-      actions: actions,
-    );
-  }
-}
-
 class DialogAction extends StatelessWidget {
   const DialogAction({
     super.key,
