+++ app/lib/common/widgets/dialog_content_text.dart
@@ -1,0 +1,14 @@
+import '../module.dart';
+
+class DialogContentText extends StatelessWidget {
+  const DialogContentText(this.text, {
+    super.key,
+  });
+
+  final String text;
+
+  @override
+  Widget build(BuildContext context) {
+    return Text(text, style: PharMeTheme.textTheme.bodyLarge);
+  }
+}
