+++ app/lib/common/widgets/full_width_button.dart
@@ -19,13 +19,10 @@ class FullWidthButton extends StatelessWidget {
       child: ElevatedButton(
         onPressed: enabled ? action : null,
         style: ButtonStyle(
-          shape: MaterialStateProperty.all<RoundedRectangleBorder>(
-            RoundedRectangleBorder(
-              borderRadius: BorderRadius.circular(32),
-            ),
-          ),
+          backgroundColor:
+            MaterialStateProperty.all<Color>(PharMeTheme.primaryColor),
         ),
-        child: Text(text),
+        child: Text(text, style: PharMeTheme.textTheme.bodyLarge),
       ),
     );
   }
