+++ app/lib/common/widgets/module.dart
@@ -1,6 +1,8 @@
-export 'adaptive_dialog.dart';
 export 'checkbox_list_tile_wrapper.dart';
 export 'context_menu.dart';
+export 'dialog_action.dart';
+export 'dialog_content_text.dart';
+export 'dialog_wrapper.dart';
 export 'drug_list/builder.dart';
 export 'drug_list/cubit.dart';
 export 'headings.dart';
